perf: Replace u32 with u8 and point update on game update

+++ src/game/board.rs
@@ -18,7 +18,7 @@ impl Board {
 
     /// Returns the number of tiles of the given color on the board.
     /// Calculates the count every time it is called, which may be inefficient for frequent calls.
-    pub fn get_colored_tiles(&self, color: &Color) -> u32 {
+    pub fn get_colored_tiles(&self, color: &Color) -> u8 {
         let mut num = 0;
         for row in self.board.iter() {
             for cell in row.iter() {
